[Decomp][Player] Extract death-knight rune system into RuneMgr
